+++ packages/engine-server/src/config.ts
@@ -1,12 +1,27 @@
 import {
+  CleanDendronSiteConfig,
   CONSTANTS,
   DendronConfig,
   DendronSiteConfig,
+  getStage
 } from "@dendronhq/common-all";
 import { readYAML, writeYAML } from "@dendronhq/common-server";
 import fs from "fs-extra";
 import _ from "lodash";
 import path from "path";
+
+export class ConfigUtils {
+  static usePrettyRef(config: DendronConfig) {
+    let usePrettyRefs: boolean | undefined = _.find(
+      [config?.usePrettyRefs, config?.site?.usePrettyRefs],
+      (ent) => !_.isUndefined(ent)
+    );
+    if (_.isUndefined(usePrettyRefs)) {
+      usePrettyRefs = true;
+    }
+    return usePrettyRefs;
+  }
+}
 
 export class DConfig {
   static configPath(configRoot: string): string {
@@ -44,10 +59,15 @@ export class DConfig {
     return config;
   }
 
+  static getSiteIndex(sconfig: DendronSiteConfig) {
+    let { siteIndex, siteHierarchies } = sconfig;
+    return siteIndex || siteHierarchies[0];
+  }
+
   /**
    * fill in defaults
    */
-  static cleanSiteConfig(config: DendronSiteConfig): DendronSiteConfig {
+  static cleanSiteConfig(config: DendronSiteConfig): CleanDendronSiteConfig {
     let out: DendronSiteConfig = _.defaults(config, {
       copyAssets: true,
       usePrettyRefs: true,
@@ -61,15 +81,29 @@ export class DConfig {
       writeStubs: true,
       description: "Personal knowledge space",
     });
-    let { siteRootDir, siteHierarchies, siteIndex } = out;
+    let { siteRootDir, siteHierarchies, siteIndex, siteUrl } = out;
+    if (process.env["SITE_URL"]) {
+      siteUrl = process.env["SITE_URL"];
+    }
     if (!siteRootDir) {
       throw `siteRootDir is undefined`;
+    }
+    if (!siteUrl && getStage() === "dev") {
+      // this gets overridden in dev so doesn't matter
+      siteUrl = "https://foo";
+    }
+    if (!siteUrl) {
+      throw `siteUrl is undefined. See https://dendron.so/notes/f2ed8639-a604-4a9d-b76c-41e205fb8713.html#siteurl for more details`;
     }
     if (_.size(siteHierarchies) < 1) {
       throw `siteHiearchies must have at least one hiearchy`;
     }
-    out.siteIndex = siteIndex || siteHierarchies[0];
-    return out;
+    siteIndex = this.getSiteIndex(config);
+    return {
+      ...out,
+      siteIndex,
+      siteUrl,
+    };
   }
 
   static writeConfig({