+++ packages/engine-server/src/markdown/remark/noteRefsV2.ts
@@ -9,7 +9,7 @@ import {
   NotePropsV2,
   NoteUtilsV2,
   RespV2,
-  VaultUtils,
+  VaultUtils
 } from "@dendronhq/common-all";
 import { file2Note } from "@dendronhq/common-server";
 import _ from "lodash";
@@ -24,7 +24,7 @@ import {
   DendronASTNode,
   DendronASTTypes,
   NoteRefNoteV4,
-  NoteRefNoteV4_LEGACY,
+  NoteRefNoteV4_LEGACY
 } from "../types";
 import { MDUtilsV4, renderFromNoteProps } from "../utils";
 import { LinkUtils } from "./utils";
@@ -460,7 +460,9 @@ function convertNoteRefHelperAST(
         anchorEndIndex
       )
     );
-    const tmpProc = MDUtilsV4.procFull({ ...procOpts });
+    let tmpProc = MDUtilsV4.procFull({ ...procOpts });
+    tmpProc = MDUtilsV4.setDendronData(tmpProc, {insideNoteRef: true})
+
     // let tmpProc = proc.data("procFull") as Processor;
     const { dest } = MDUtilsV4.getDendronData(tmpProc);
     if (dest === DendronASTDest.HTML) {
@@ -604,3 +606,4 @@ function renderPrettyAST(opts: {
 
 export { plugin as noteRefsV2 };
 export { PluginOpts as NoteRefsOptsV2 };
+