+++ packages/engine-server/src/markdown/remark/wikiLinks.ts
@@ -154,13 +154,18 @@ function attachParser(proc: Unified.Processor) {
         vname: out.vaultName,
       });
       if (_.isUndefined(maybeVault)) {
-        throw new DendronError({
-          msg: `fname: ${fname}, vault ${
-            out.vaultName
-          } not found in ${JSON.stringify(out)}`,
-        });
+        addError(
+          proc,
+          new DendronError({
+            msg: `fname: ${fname}, vault ${
+              out.vaultName
+            } not found in ${JSON.stringify(out)}`,
+          })
+        );
+      } else {
+        vault = maybeVault;
       }
-      vault = maybeVault;
+      // default to current note
     }
     if (
       dest !== DendronASTDest.MD_DENDRON &&