+++ packages/engine-test-utils/src/__tests__/engine-server/markdown/utils.ts
@@ -1,4 +1,8 @@
-import { DEngineClientV2, DVault } from "@dendronhq/common-all";
+import {
+  DEngineClientV2,
+  DuplicateNoteAction,
+  DVault,
+} from "@dendronhq/common-all";
 import { AssertUtils, TestPresetEntryV4 } from "@dendronhq/common-test-utils";
 import { DendronASTDest, MDUtilsV4, VFile } from "@dendronhq/engine-server";
 import _ from "lodash";
@@ -9,6 +13,16 @@ export async function checkVFile(resp: VFile, ...match: string[]) {
     await AssertUtils.assertInString({
       body: resp.toString(),
       match,
+    })
+  ).toBeTruthy();
+}
+
+export async function checkNotInVFile(resp: VFile, ...nomatch: string[]) {
+  expect(resp).toMatchSnapshot();
+  expect(
+    await AssertUtils.assertInString({
+      body: resp.toString(),
+      nomatch,
     })
   ).toBeTruthy();
 }
@@ -84,4 +98,20 @@ export const createProcTests = (opts: {
       .filter((ent) => !_.isUndefined(ent));
   }
   return allTests;
+};
+
+export const dupNote = (payload: DVault | string[]) => {
+  const out: any = {
+    duplicateNoteBehavior: {
+      action: DuplicateNoteAction.USE_VAULT,
+    },
+  };
+  if (_.isArray(payload)) {
+    out.duplicateNoteBehavior.payload = payload;
+  } else {
+    out.duplicateNoteBehavior.payload = {
+      vault: payload,
+    };
+  }
+  return out;
 };