+++ packages/plugin-core/src/commands/index.ts
@@ -3,12 +3,14 @@ import { CodeCommandConstructor } from "./base";
 import { ChangeWorkspaceCommand } from "./ChangeWorkspace";
 import { ConfigureCommand } from "./ConfigureCommand";
 import { ConfigurePodCommand } from "./ConfigurePodCommand";
+import { ConfigureWithUICommand } from "./ConfigureWithUI";
 import { ContributeCommand } from "./Contribute";
 import { CopyNoteLinkCommand } from "./CopyNoteLink";
 import { CopyNoteRefCommand } from "./CopyNoteRef";
 import { CopyNoteURLCommand } from "./CopyNoteURL";
 import { CreateDailyJournalCommand } from "./CreateDailyJournal";
 import { DeleteNodeCommand } from "./DeleteNodeCommand";
+import { DiagnosticsReportCommand } from "./DiagnosticsReport";
 import { DoctorCommand } from "./Doctor";
 import { DumpStateCommand } from "./DumpStateCommand";
 import { ExportPodCommand } from "./ExportPod";
@@ -16,7 +18,9 @@ import { GoDownCommand } from "./GoDownCommand";
 import { GotoNoteCommand } from "./GotoNote";
 import { GoUpCommand } from "./GoUpCommand";
 import { ImportPodCommand } from "./ImportPod";
+import { MoveNoteCommand } from "./MoveNoteCommand";
 import { OpenLogsCommand } from "./OpenLogs";
+import { PasteLinkCommand } from "./PasteLink";
 import { PublishPodCommand } from "./PublishPod";
 import { RefactorHierarchyCommandV2 } from "./RefactorHierarchyV2";
 import { RenameNoteV2aCommand } from "./RenameNoteV2a";
@@ -25,17 +29,14 @@ import { RestoreVaultCommand } from "./RestoreVault";
 import { SetupWorkspaceCommand } from "./SetupWorkspace";
 import { ShowHelpCommand } from "./ShowHelp";
 import { ShowPreviewCommand } from "./ShowPreview";
+import { SignInCommand } from "./SignIn";
+import { SignUpCommand } from "./SignUp";
+import { SiteBuildCommand } from "./SiteBuild";
+import { SitePreviewCommand } from "./SitePreview";
 import { SnapshotVaultCommand } from "./SnapshotVault";
-import { ConfigureWithUICommand } from "./ConfigureWithUI";
 import { UpgradeSettingsCommand } from "./UpgradeSettings";
 import { VaultAddCommand } from "./VaultAddCommand";
 import { VaultRemoveCommand } from "./VaultRemoveCommand";
-import { SiteBuildCommand } from "./SiteBuild";
-import { SitePreviewCommand } from "./SitePreview";
-import { SignUpCommand } from "./SignUp";
-import { SignInCommand } from "./SignIn";
-import { MoveNoteCommand } from "./MoveNoteCommand";
-import { DiagnosticsReportCommand } from "./DiagnosticsReport";
 
 const ALL_COMMANDS = [
   ArchiveHierarchyCommand,
@@ -59,6 +60,7 @@ const ALL_COMMANDS = [
   ImportPodCommand,
   //   LookupCommand,
   OpenLogsCommand,
+  PasteLinkCommand,
   PublishPodCommand,
   MoveNoteCommand,
   RefactorHierarchyCommandV2,